Add A-Z index
Reuses common SearchReferencesController (renamed to
SearchReferencesBaseController) to return list of search
references/synonyms.

Does no pagination, but requires letter to be passed.
