+++ app/controllers/api/v1/chapters/search_references_controller.rb
@@ -1,7 +1,7 @@
 module Api
   module V1
     module Chapters
-      class SearchReferencesController < Api::V1::SearchReferencesController
+      class SearchReferencesController < Api::V1::SearchReferencesBaseController
         private
 
         def search_reference_collection
