+++ app/views/api/v1/search_references_base/_heading.json.rabl
