+++ app/views/api/v1/search_references_base/_section.json.rabl
