+++ app/views/api/v1/search_references_base/create.json.rabl
@@ -1,0 +1,7 @@
+object @search_reference
+
+attributes :id, :title, :referenced_entity
+
+node(false) { |search_reference|
+  partial("api/v1/search_references_base/#{search_reference.referenced_entity.class.name.underscore}", object: :referenced_entity)
+}
