+++ app/views/api/v1/search_references_base/show.json.rabl
